Change rating type to BigDecimal instead of Integer

+++ backend/mediaarchive/src/main/java/com/isaiah/mediaarchive/model/entity/UserMediaEntity.java
@@ -32,7 +32,7 @@ public class UserMediaEntity {
     @Enumerated(EnumType.STRING)
     private CollectionStatusEnum status;
 
-    private Integer personalRating;
+    private BigDecimal personalRating;
 
     @Column(length = 2000)
     private String notes;
